requirements.txt
aiohttp
disnake
python-dotenv
colorlog